requirements.txt
ultralytics>=8.0.0
streamlit>=1.33.0
gTTS>=2.5.1
Pillow>=10.2.0
numpy>=1.26.0
opencv-python-headless>=4.8.0
roboflow
